writing a bit of doc, moving crawler together
