+++ AUTHORS.rst
@@ -1,6 +1,8 @@
-pyAggr3g470r is a free software written and maintained
+Original author
+---------------
+JARR is a free software written and maintained
 by Cédric Bonhomme https://www.cedricbonhomme.org
 
 Contributors
-````````````
+------------
 - François Schmidts http://1pxsolidblack.pl/
